+++ Sources/DTTableViewManager/DTTableViewDelegate.swift
@@ -502,4 +502,22 @@ open class DTTableViewDelegate : DTTableViewDelegateWrapper, UITableViewDelegate
         return tableView.selectionFollowsFocus
     }
     #endif
+    
+#if swift(>=5.7) || (os(macOS) && swift(>=5.7.1)) // Xcode 14.0 AND macCatalyst on Xcode 14.1 (which will have swift> 5.7.1)
+    @available(iOS 16, tvOS 16, *)
+    /// Implementation for `UITableViewDelegate` protocol
+    public func tableView(_ tableView: UITableView, canPerformPrimaryActionForRowAt indexPath: IndexPath) -> Bool {
+        if let canPerform = performCellReaction(.canPerformActionForRowAtIndexPath, location: indexPath, provideCell: true) as? Bool {
+            return canPerform
+        }
+        return (delegate as? UITableViewDelegate)?.tableView?(tableView, canPerformPrimaryActionForRowAt: indexPath) ?? false
+    }
+    
+    @available(iOS 16, tvOS 16, *)
+    /// Implementation for `UITableViewDelegate` protocol
+    public func tableView(_ tableView: UITableView, performPrimaryActionForRowAt indexPath: IndexPath) {
+        _ = performCellReaction(.performPrimaryActionForRowAtIndexPath, location: indexPath, provideCell: true)
+        (delegate as? UITableViewDelegate)?.tableView?(tableView, performPrimaryActionForRowAt: indexPath)
+    }
+#endif
 }